fix(diff): then/else are branch tokens too; label logic moves to diff_label.go (Codex P2)
An if action with the same service in then and else no longer collapses
into indistinguishable rows. The label/anchor half of the diff renderer
moves into cli/diff_label.go, bringing diff_format.go back under the
~400-line guardrail (321 + 140).
